FullScreen fixed

+++ src/Boid.java
@@ -76,7 +76,7 @@ public class Boid {
     
         double headingAngle = Math.atan2(velocity.y, velocity.x);  // direction in which the boid is currently heading
         
-        g2d.translate((int)(position.x),(int)(position.y));
+        g2d.translate(position.x,position.y);
         g2d.rotate(headingAngle);
         
         // Drawing the boid here
